+++ src/connection/UniqueConstraintViolationError.ts
@@ -1,0 +1,6 @@
+export class UniqueConstraintViolationError extends Error {
+  constructor(options: ErrorOptions = {}) {
+    super("A unique constraint was violated.", options);
+    this.name = "UniqueConstraintViolationError";
+  }
+}